requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
gymnasium>=0.28.0
opencv-python>=4.7.0

# Configuration and logging
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb>=0.15.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Development tools
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0

# Optional: Additional vision models
# timm>=0.9.0  # For more vision backbones
# albumentations>=1.3.0  # For advanced image augmentations